Split out export code from backups.py
